% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{Accuracy}
\alias{Accuracy}
\title{Accuracy}
\usage{
Accuracy(py, y, s = frequency(y), collectFun = mean)
}
\arguments{
\item{py}{matrix of forecasts (h x nMethods x nForecasts)}

\item{y}{a matrix of actual values (n x nForecasts)}

\item{s}{seasonal period, number of observations per year}

\item{collectFun}{aggregation function (mean, median, etc.)}
}
\value{
Table of results
}
\description{
Accuracy for 1 time series y and several forecasting
 methods py and h steps ahead py is h x nMethods x nSeries
}
\examples{
\dontrun{Accuracy(py, y, 12)}
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
