% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{acft}
\alias{acft}
\title{acft}
\usage{
acft(MApoly = 1, ARpoly = 1, ncoef = 38, s = 1)
}
\arguments{
\item{MApoly}{coefficients of numerator polynomial in descending order}

\item{ARpoly}{coefficients of denominator polynomial in descending order}

\item{ncoef}{number of coefficients}

\item{s}{seasonal period, number of observations per year}
}
\value{
Theoretical autocorrelation functions
}
\description{
Theoretical autocorrelation functions of ARMA models
}
\examples{
acft(c(1, -0.8), c(1, 0.8))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
