% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airpas}
\alias{airpas}
\title{Airpassengers in Spain}
\format{
Time series objects.
\describe{
  Monthly data from 1969
  
  https://portal.mineco.gob.es/es-es/economiayempresa/EconomiaInformesMacro/Paginas/bdsice.aspx
}
}
\usage{
airpas
}
\description{
Foreign arrivals by air in Spain in thousands of passengers (airpas).
}
\examples{
\dontrun{
 airpas
}
}
\keyword{datasets}
