% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIMAmodel.R
\name{ARIMAsetup}
\alias{ARIMAsetup}
\title{ARIMAsetup}
\usage{
ARIMAsetup(
  y,
  u = NULL,
  model = NULL,
  cnst = NULL,
  s = frequency(y),
  criterion = "bic",
  h = 2 * s,
  verbose = FALSE,
  lambda = 1,
  maxOrders = c(3, 2, 3, 2, 1, 2),
  bootstrap = FALSE,
  nSimul = 5000,
  fast = FALSE
)
}
\arguments{
\item{y}{a time series to forecast (it may be either a numerical vector or
a time series object). This is the only input required. If a vector, the additional
input \code{s} should be supplied compulsorily (see below).}

\item{u}{a matrix of input time series. If 
the output wanted to be forecast, matrix \code{u} should contain future values for inputs.}

\item{model}{the model to estimate. A vector c(p,d,q,P,D,Q) containing the model orders
of an ARIMA(p,d,q)x(P,D,Q)_s model. A constant may be estimated with the
cnst input.
Use a NULL to automatically identify the ARIMA model.}

\item{cnst}{flag to include a constant in the model (TRUE/FALSE/NULL). Use NULL to estimate}

\item{s}{seasonal period of time series (1 for annual, 4 for quarterly, ...)}

\item{criterion}{information criterion for identification stage ("aic", "bic", "aicc")}

\item{h}{forecast horizon. If the model includes inputs h is not used, the lenght of u is used instead.}

\item{verbose}{intermediate estimation output (TRUE / FALSE)}

\item{lambda}{Box-Cox lambda parameter (NULL: estimate)}

\item{maxOrders}{a vector c(p,d,q,P,D,Q) containing the maximum orders of model orders 
to search for in the automatic identification}

\item{bootstrap}{use bootstrap simulation for predictive distributions}

\item{nSimul}{number of simulation runs for bootstrap simulation of predictive distributions}

\item{fast}{fast identification (avoids post-identification checks)}
}
\value{
An object of class \code{ARIMA}. See \code{ARIMAforecast}.
}
\description{
Sets up ARIMA general models
}
\details{
See help of \code{ARIMAforecast}.
}
\examples{
\dontrun{
y <- log(AirPAssengers)
m1 <- ARIMAsetup(y)
m1 <- ARIMAsetup(y, lambda = NULL)
}
}
\seealso{
\code{\link{ARIMA}}, \code{\link{ARIMAforecast}}, \code{\link{ARIMAvalidate}},
}
\author{
Diego J. Pedregal
}
