% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIMAvalidate.R
\name{ARIMAvalidate}
\alias{ARIMAvalidate}
\title{ARIMAvalidate}
\usage{
ARIMAvalidate(m)
}
\arguments{
\item{m}{an object of type \code{ARIMA} created with \code{ARIMAmodel}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for ARIMA models
}
\examples{
\dontrun{
m1 <- ARIMAmodel(log(gdp))
m1 <- ARIMAvalidate(m1)
}
}
\seealso{
\code{\link{ARIMA}}, \code{\link{ARIMAmodel}}, \code{\link{ARIMAvalidate}},
}
\author{
Diego J. Pedregal
}
