% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{arma2tsi}
\alias{arma2tsi}
\title{arma2tsi}
\usage{
arma2tsi(MApoly, ARpoly, n = 100)
}
\arguments{
\item{MApoly}{coefficients of numerator polynomial in descending order}

\item{ARpoly}{coefficients of denominator polynomial in descending order}

\item{n}{number of coefficients}
}
\description{
AR polynomial coefficients of ARMA model
}
\author{
Diego J. Pedregal
}
