% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{armaFilter}
\alias{armaFilter}
\title{armaFilter}
\usage{
armaFilter(MA = 1, AR = 1, y)
}
\arguments{
\item{MA}{numerator polynomial}

\item{AR}{denominator polynomial}

\item{y}{a vector, ts or tsibble object}
}
\value{
Filtered time series
}
\description{
Filter of time series
}
\examples{
y <- armaFilter(1, c(1 , -0.8), rnorm(200))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
