% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSS3functions.R
\name{auxInvBoxCox}
\alias{auxInvBoxCox}
\title{auxInvBoxCox}
\usage{
auxInvBoxCox(y, lambda)
}
\arguments{
\item{y}{matrix, array or vector}

\item{lambda}{lambda parameter of Box-Cox transformation}
}
\description{
Inverse of Box-Cox transformation
}
\author{
Diego J. Pedregal
}
