% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{box.cox}
\alias{box.cox}
\title{box.cox}
\usage{
box.cox(x, lambda)
}
\arguments{
\item{x}{Time series object.}

\item{lambda}{Lambda parameter for Box-Cox transform.}
}
\description{
Runs Box-Cox transform of a time series
}
\examples{
\dontrun{
y <- box.cox(AirPassengers, 0.5)
plot(y)
}
}
\seealso{
\code{\link{inv.box.cox}}, \code{\link{UC}}, \code{\link{UCforecast}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
