% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch4}
\alias{ch4}
\title{Methane concentration at Cape Grim in Australia (ch4).}
\format{
Time series objects.
\describe{
  Monthly data from January 1992 to December 2019
}
}
\source{
\href{https://www.csiro.au/en/Research/OandA/Areas/Assessing-our-climate/Latest-greenhouse-gas-data}{CH4 data}
}
\usage{
ch4
}
\description{
Methane concentration at Cape Grim in Australia (ch4).
}
\examples{
\dontrun{
 ch4
}
}
\keyword{datasets}
