% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R
\name{coef.UComp}
\alias{coef.UComp}
\title{coef.UComp}
\usage{
\method{coef}{UComp}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{UComp}.}

\item{...}{Additional inputs to function.}
}
\description{
Extract model coefficients of UComp object
}
\details{
See help of \code{UC}.
}
\examples{
y <- log(AirPassengers)
m1 <- UCmodel(y, model = "llt/equal/arma(0,0)")
coef(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
