% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{colMedians}
\alias{colMedians}
\title{colMedians}
\usage{
colMedians(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a matrix}

\item{na.rm}{boolean indicating whether to remove nans}

\item{...}{rest of inputs}
}
\value{
A vector with all the medians
}
\description{
Medians of matrix by columns
}
\examples{
s <- colMedians(matrix(4, 3, 2))
}
\seealso{
\code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
