% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{conv}
\alias{conv}
\title{conv}
\usage{
conv(...)
}
\arguments{
\item{...}{list of vectors to convolute}
}
\value{
Convolution of all input vectors
}
\description{
1D convolution: filtering or polynomial multiplication
}
\examples{
conv(c(1, -1), c(1, -2, 1))
conv(c(1, -1), c(1, 0.8))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
