% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cusum}
\alias{cusum}
\title{cusum}
\usage{
cusum(y, runFromTest = FALSE)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{runFromTest}{internal check variable}
}
\description{
Cusum and cusumsq tests
}
\examples{
cusum(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
