% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidFunctions.R
\name{DAYforecast}
\alias{DAYforecast}
\title{DAYforecast}
\usage{
DAYforecast(
  y,
  u = NULL,
  initial_date,
  h = 24,
  lambda = 1,
  criterion = "aic",
  p0 = -9999.9,
  verbose = FALSE
)
}
\arguments{
\item{y}{A numeric vector, \code{ts} object or list containing daily observations of the
dependent variable.}

\item{u}{Optional matrix or vector of exogenous regressors. If provided, it must cover
both the estimation and forecast periods.}

\item{initial_date}{A character or \code{Date} object indicating the calendar start date
of the series.}

\item{h}{Integer indicating the forecast horizon (number of days ahead).}

\item{lambda}{Box–Cox transformation parameter (NULL for automatic identification).}

\item{criterion}{Information criterion used for model selection (AIC, BIC, AICc).}

\item{p0}{Initial parameter vector for Maximum Likelihood estimation.}

\item{verbose}{if \code{TRUE}, prints estimation details.}
}
\value{
An object of class \code{UComp}. It is a list containing the model specification,
        estimated parameters, and forecasting results.
}
\description{
Estimates and forecasts daily time series with weekly and monthly seasonal patterns
}
\details{
\code{DAYforecast} estimates and forecasts daily univariate time series using
Dynamic Harmonic Regression (DHR).

The function is tailored for daily data and explicitly accounts for:
\itemize{
\item Monthly seasonality through sinusoidal terms based on calendar months
\item Weekly seasonality through sinusoidal terms with periods of 7 days
}

Standard methods applicable to \code{UComp} objects are available.
}
\examples{
\donttest{
y <- rnorm(400)
m <- DAYforecast(y, initial_date = "2015-01-01", h = 30)
}

}
\seealso{
\code{\link{UC}}, \code{\link{UCforecast}}, \code{\link{UCfilter}},
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}},
         \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
