% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dif}
\alias{dif}
\title{dif}
\usage{
dif(y, difs = 1, seas = 1)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{difs}{vector with differencing orders}

\item{seas}{vector of seasonal periods}
}
\value{
Differenced time series
}
\description{
Discrete differencing of time series
}
\examples{
dif(AirPassengers)
dif(AirPassengers, 2)
dif(AirPassengers, c(1, 1), c(1, 12))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
