% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETScomponents.R
\name{ETScomponents}
\alias{ETScomponents}
\title{ETScomponents}
\usage{
ETScomponents(m)
}
\arguments{
\item{m}{an object of type \code{ETS} created with \code{ETSforecast}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{comp}{Estimated components in matrix form}
}
\description{
Estimates components of ETS models
}
\examples{
\dontrun{
m1 <- ETS(log(gdp))
m1 <- ETScomponents(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSforecast}}, \code{\link{ETSvalidate}},
         \code{\link{ETSestim}}
}
\author{
Diego J. Pedregal
}
