% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSvalidate.R
\name{ETSvalidate}
\alias{ETSvalidate}
\title{ETSvalidate}
\usage{
ETSvalidate(m)
}
\arguments{
\item{m}{an object of type \code{ETS} created with \code{ETSforecast}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for ETS models
}
\examples{
\dontrun{
m1 <- ETSforecast(log(gdp))
m1 <- ETSvalidate(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSforecast}}, \code{\link{ETSvalidate}},
         \code{\link{ETScomponents}}
}
\author{
Diego J. Pedregal
}
