% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extract}
\alias{extract}
\title{extract}
\usage{
extract(x, col, accordingTo = 1)
}
\arguments{
\item{x}{a data frame}

\item{col}{column to be ordered}

\item{accordingTo}{column to take as the pattern}
}
\value{
data frame reordered
}
\description{
Reorder data frame returning column col reordered
according to the values in column accordingTo
}
\author{
Diego J. Pedregal
}
