% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gaussTest}
\alias{gaussTest}
\title{gaussTest}
\usage{
gaussTest(y, runFromTests = FALSE)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{runFromTests}{internal check}
}
\description{
Gaussianity tests
}
\examples{
gaussTest(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
