% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Spanish GDP}
\format{
Time series objects.
\describe{
  Quarterly since 1995
  
  https://portal.mineco.gob.es/es-es/economiayempresa/EconomiaInformesMacro/Paginas/bdsice.aspx
}
}
\usage{
gdp
}
\value{
No return value, called for side effects
}
\description{
Quarterly real Spanish Gross Domestic Product (gdp)
}
\examples{
\donttest{
 gdp
}
}
\keyword{datasets}
