% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ident}
\alias{ident}
\title{ident}
\usage{
ident(y, nCoef = min(37, floor(length(y)/4)), nPar = 0, runFromTests = FALSE)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{nCoef}{number of autocorrelation coefficients to estimate}

\item{nPar}{number of parameters in a model if y is a residual}

\item{runFromTests}{internal check}
}
\value{
A vector with all the dimensions
}
\description{
Autocorrelation functions of a time series
}
\examples{
ident(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
