% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{inv.box.cox}
\alias{inv.box.cox}
\title{inv.box.cox}
\usage{
inv.box.cox(x, lambda)
}
\arguments{
\item{x}{Transformed time series object.}

\item{lambda}{Lambda parameter used for Box-Cox transform.}
}
\description{
Runs inverse of Box-Cox transform of a time series
}
\examples{
\dontrun{
y <- inv.box.cox(box.cox(AirPassengers, 0.5), 0.5)
plot(y)
}
}
\seealso{
\code{\link{box.cox}}, \code{\link{UC}}, \code{\link{UCforecast}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
