% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSS3functions.R
\name{invBoxCox}
\alias{invBoxCox}
\title{invBoxCox}
\usage{
invBoxCox(y, yVar, lambda, const = 2)
}
\arguments{
\item{y}{matrix, array or vector}

\item{yVar}{matrix, array or vector of variances of y}

\item{lambda}{lambda parameter of Box-Cox transformation}

\item{const}{number of standard error for confidence band}
}
\description{
Calculates inverse of Box-Cox transformation 
with confidence bands, calculated as const time the standard error
}
\author{
Diego J. Pedregal
}
