% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipi}
\alias{ipi}
\title{Spanish Industrial Production Index}
\format{
Objeto time series.
\describe{
  Monthly since 1975
  
  https://portal.mineco.gob.es/es-es/economiayempresa/EconomiaInformesMacro/Paginas/bdsice.aspx
}
}
\usage{
ipi
}
\value{
No return value, called for side effects
}
\description{
Spanish Industrial Production Index (ipi).
}
\examples{
\donttest{
 ipi
}
}
\keyword{datasets}
