% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{modelUC2arma}
\alias{modelUC2arma}
\title{modelUC2arma}
\usage{
modelUC2arma(model)
}
\arguments{
\item{model}{a UC model}
}
\value{
arma orders
}
\description{
Extracts arma part of modelUC model
}
\author{
Diego J. Pedregal
}
