% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{modelUC2PTS}
\alias{modelUC2PTS}
\title{modelUC2PTS}
\usage{
modelUC2PTS(modelUC)
}
\arguments{
\item{modelUC}{a UC model}
}
\value{
a PTS model
}
\description{
Translates modelUC to model PTS
}
\author{
Diego J. Pedregal
}
