% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OECDgdp}
\alias{OECDgdp}
\title{OECD GDP}
\format{
Time series objects.
\describe{
  Quarterly data from 1962 to 2019
  
  https://portal.mineco.gob.es/es-es/economiayempresa/EconomiaInformesMacro/Paginas/bdsice.aspx
}
}
\usage{
OECDgdp
}
\value{
No return value, called for side effects
}
\description{
Seasonally adjusted quarterly OECD real gross domestic product (OECDgdp).
}
\examples{
\donttest{
 OECDgdp
}
}
\keyword{datasets}
