% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R
\name{plot.UComp}
\alias{plot.UComp}
\title{plot.UComp}
\usage{
\method{plot}{UComp}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{UComp}.}

\item{...}{Additional inputs to function.}
}
\description{
Plot components of UComp object
}
\details{
See help of \code{UC}.
}
\examples{
y <- log(AirPassengers)
m1 <- UCmodel(y, model = "llt/equal/arma(0,0)")
plot(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
