% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotAcfPacf}
\alias{plotAcfPacf}
\title{plotAcfPacf}
\usage{
plotAcfPacf(ACF, PACF, s = 1, n = NA, runFromTest = FALSE)
}
\arguments{
\item{ACF}{variable to plot}

\item{PACF}{second variable to plot}

\item{s}{seasonal period}

\item{n}{number of coefficients}

\item{runFromTest}{internal check variable}
}
\description{
Plot of ACF and PACF
}
\author{
Diego J. Pedregal
}
