% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotBar}
\alias{plotBar}
\title{plotBar}
\usage{
plotBar(ACF, s = 1, n = NA, label = "ACF")
}
\arguments{
\item{ACF}{variable to plot}

\item{s}{seasonal period}

\item{n}{number of coefficients}

\item{label}{label for plot}
}
\value{
Handle of plot
}
\description{
Plot variable in bars
}
\author{
Diego J. Pedregal
}
