% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plotSlide}
\alias{plotSlide}
\title{plotSlide}
\usage{
plotSlide(py1, y, orig, step = 1, errorFun, collectFun = mean)
}
\arguments{
\item{py1}{output from slide function}

\item{y}{a vector, matrix or list of time series (the same used in slide call)}

\item{orig}{starting forecasting origin (the same used in slide call)}

\item{step}{observations ahead to move the forecasting origin (the same used in slide call)}

\item{errorFun}{user function to calculate error measures}

\item{collectFun}{aggregation function (mean, median, etc.)}
}
\value{
An array of forecasting errors of dimensions (horizon x nOrigs x nModels x nSeries)
}
\description{
Plot summarised results from slide
}
\examples{
\dontrun{plotSlide(py1, AirPassengers, 100, 1, errorFun)}
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
