% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plus_one}
\alias{plus_one}
\title{plus_one}
\usage{
plus_one(y)
}
\arguments{
\item{y}{a ts object}
}
\value{
Next time stamp
}
\description{
Returns date of next to end time series y
}
\author{
Diego J. Pedregal
}
