% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R
\name{predict.UComp}
\alias{predict.UComp}
\title{predict.UComp}
\usage{
\method{predict}{UComp}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{UComp}.}

\item{...}{Additional inputs to function.}
}
\value{
A list with components \code{pred} for the predictions and 
        \code{se} for standard errors (if \code{se.fit = TRUE})
}
\description{
Forecasting using structural Unobseved Components models
}
\details{
See help of \code{UC}.
}
\examples{
y <- log(AirPassengers)
m1 <- UCmodel(y, model = "llt/eq/arma(0,0)")
f1 <- predict(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
