% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{PTS2modelUC}
\alias{PTS2modelUC}
\title{PTS2modelUC}
\usage{
PTS2modelUC(model, armaOrders = c(0, 0))
}
\arguments{
\item{model}{a PTS model}

\item{armaOrders}{arma orders of noise model}
}
\value{
a UC model
}
\description{
Translates PTS model to UC model
}
\author{
Diego J. Pedregal
}
