% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{PTScomponents}
\alias{PTScomponents}
\title{PTScomponents}
\usage{
PTScomponents(m)
}
\arguments{
\item{m}{an object of type \code{PTS} created with \code{PTSmodel}}
}
\value{
The same input object with the appropriate fields
filled in, in particular:
\itemize{
\item comp:  Estimated components in matrix form
}
}
\description{
Estimates components of PTS models
}
\examples{
\dontrun{
m1 <- PTS(log(AirPassengers))
m1 <- PTScomponents(m1)
}
}
\seealso{
\code{\link{PTSmodel}}, \code{\link{PTSsetup}}, \code{\link{PTSestim}},
         \code{\link{PTSvalidate}}, \code{\link{PTS}}
}
\author{
Diego J. Pedregal
}
