% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{PTSvalidate}
\alias{PTSvalidate}
\title{PTSvalidate}
\usage{
PTSvalidate(m, verbose = TRUE)
}
\arguments{
\item{m}{an object of type \code{PTS} created with \code{PTSmodel}}

\item{verbose}{verbose mode TRUE/FALSE}
}
\value{
The same input object with the appropriate fields
filled in, in particular:
\itemize{
\item table: Estimation and validation table
}
}
\description{
Shows a table of estimation and diagnostics results for PTS models
}
\examples{
\dontrun{
m1 <- PTSmodel(log(AirPassengers))
m1 <- PTSvalidate(m1)
}
}
\seealso{
\code{\link{PTSmodel}}, \code{\link{PTSsetup}}, \code{\link{PTSestim}},
         \code{\link{PTScomponents}}, \code{\link{PTS}}
}
\author{
Diego J. Pedregal
}
