% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{removeNaNs}
\alias{removeNaNs}
\title{removeNaNs}
\usage{
removeNaNs(x)
}
\arguments{
\item{x}{a vector or a ts object}
}
\value{
vector with nans removed (only those at beginning or end)
}
\description{
Remove nans at beginning or end of vector
}
\author{
Diego J. Pedregal
}
