% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales index for large retailers in Spain}
\format{
Time series objects.
\describe{
  Monthly data from January 1995 to December 2019
  
  https://portal.mineco.gob.es/es-es/economiayempresa/EconomiaInformesMacro/Paginas/bdsice.aspx
}
}
\usage{
sales
}
\description{
Sales index for food of large retailers in Spain
}
\examples{
\dontrun{
 sales
}
}
\keyword{datasets}
