% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R, R/tools.R
\name{size}
\alias{size}
\title{size}
\usage{
size(y)

size(y)
}
\arguments{
\item{y}{a vector, matrix or array}
}
\value{
A vector with all the dimensions
}
\description{
size of vectors or matrices

Size of vector, matrix or array
}
\examples{
s <- size(matrix(4, 3, 2))
s <- size(rep(4, 3))
s <- size(array(4, c(3, 2, 2)))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}}
}
\author{
Diego J. Pedregal
}
