% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{slideAux}
\alias{slideAux}
\title{slideAux}
\usage{
slideAux(
  y,
  orig,
  forecFun,
  h = 12,
  step = 1,
  output = TRUE,
  graph = TRUE,
  window = NA,
  parallel = FALSE,
  isList = FALSE,
  ...
)
}
\arguments{
\item{y}{a vector or matrix of time series}

\item{orig}{starting forecasting origin}

\item{forecFun}{user function that implements forecasting methods}

\item{h}{forecasting horizon}

\item{step}{observations ahead to move the forecasting origin}

\item{output}{output TRUE/FALSE}

\item{graph}{fraphical output TRUE/FALSE}

\item{window}{fixed window width in number of observations (NA for non fixed)}

\item{parallel}{run forecasts in parallel}

\item{isList}{whether the input data y is a list or a matrix}

\item{...}{rest of inputs to forecFun function}
}
\value{
Auxiliary output of slide function for just one time series
}
\description{
Auxiliary function run from slide
}
\author{
Diego J. Pedregal
}
