% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARIMAS3functions.R
\name{summary.ARIMA}
\alias{summary.ARIMA}
\title{summary.ARIMA}
\usage{
\method{summary}{ARIMA}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{ARIMA}.}

\item{...}{Additional inputs to function.}
}
\description{
Prints an ARIMA object on screen
}
\details{
See help of \code{ARIMA}.
}
\examples{
\dontrun{
m1 <- ARIMAforecast(log(gdp))
summary(m1)
}
}
\seealso{
\code{\link{ARIMA}}, \code{\link{ARIMAforecast}}, \code{\link{ARIMAvalidate}},
}
\author{
Diego J. Pedregal
}
