% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSS3functions.R
\name{summary.ETS}
\alias{summary.ETS}
\title{summary.ETS}
\usage{
\method{summary}{ETS}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{ETS}.}

\item{...}{Additional inputs to function.}
}
\description{
Prints an ETS object on screen
}
\details{
See help of \code{ETS}.
}
\examples{
\dontrun{
m1 <- ETSforecast(log(gdp))
summary(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSforecast}}, \code{\link{ETSvalidate}},
         \code{\link{ETScomponents}}, \code{\link{ETSestim}}
}
\author{
Diego J. Pedregal
}
