% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSS3functions.R
\name{summary.PTS}
\alias{summary.PTS}
\title{summary.PTS}
\usage{
\method{summary}{PTS}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to function.}
}
\description{
Prints an PTS object on screen
}
\details{
See help of \code{PTS}.
}
\examples{
\dontrun{
m1 <- PTSmodel(log(AirPassengers))
summary(m1)
}
}
\seealso{
\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
