% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TETSS3functions.R
\name{summary.TETS}
\alias{summary.TETS}
\title{summary.TETS}
\usage{
\method{summary}{TETS}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{TETS}.}

\item{...}{Additional inputs to function.}
}
\description{
Prints a TOBIT TETS object on screen
}
\details{
See help of \code{TETS}.
}
\examples{
\dontrun{
m1 <- TETSforecast(log(gdp))
summary(m1)
}
}
\seealso{
\code{\link{TETS}}, \code{\link{TETSforecast}}, \code{\link{TETSvalidate}},
         \code{\link{TETScomponents}}, \code{\link{TETSestim}}
}
\author{
Diego J. Pedregal
}
