% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R
\name{summary.UComp}
\alias{summary.UComp}
\title{summary.UComp}
\usage{
\method{summary}{UComp}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{UComp}.}

\item{...}{Additional inputs to function.}
}
\description{
Prints an UComp object on screen
}
\details{
See help of \code{UC}.
}
\examples{
y <- log(AirPassengers)
m1 <- UCmodel(y, model = "llt/equal/arma(0,0)")
summary(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
