% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{sumStats}
\alias{sumStats}
\title{sumStats}
\usage{
sumStats(y, decimals = 5)
}
\arguments{
\item{y}{a vector, matrix of time series}

\item{decimals}{number of decimals for table}
}
\value{
Table of values
}
\description{
Summary statistics of a matrix of variables
}
\details{
Position, dispersion, skewness, kurtosis, etc.
}
\examples{
s <- sumStats(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
