% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TETScomponents.R
\name{TETScomponents}
\alias{TETScomponents}
\title{TETScomponents}
\usage{
TETScomponents(m)
}
\arguments{
\item{m}{an object of type \code{TETS} created with \code{TETSforecast}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{comp}{Estimated components in matrix form}
}
\description{
Estimates components of TOBIT TETS models
}
\examples{
\dontrun{
m1 <- TETS(log(gdp))
m1 <- TETScomponents(m1)
}
}
\seealso{
\code{\link{TETS}}, \code{\link{TETSforecast}}, \code{\link{TETSvalidate}},
         \code{\link{TETSestim}}
}
\author{
Diego J. Pedregal
}
