% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TETSvalidate.R
\name{TETSvalidate}
\alias{TETSvalidate}
\title{TETSvalidate}
\usage{
TETSvalidate(m)
}
\arguments{
\item{m}{an object of type \code{TETS} created with \code{TETSforecast}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for TOBIT TETS models
}
\examples{
\dontrun{
m1 <- TETSforecast(log(gdp))
m1 <- TETSvalidate(m1)
}
}
\seealso{
\code{\link{TETS}}, \code{\link{TETSforecast}}, \code{\link{TETSvalidate}},
         \code{\link{TETScomponents}}
}
\author{
Diego J. Pedregal
}
