% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCS3functions.R
\name{tsdiag.UComp}
\alias{tsdiag.UComp}
\title{tsdiag.UComp}
\usage{
\method{tsdiag}{UComp}(object, gof.lag = NULL, ...)
}
\arguments{
\item{object}{Object of class \dQuote{UComp}.}

\item{gof.lag}{Maximum number of lags for pormanteau Ljung-Box test}

\item{...}{Additional inputs to function.}
}
\description{
Diagnostic plots for UComp objects
}
\details{
See help of \code{UC}.
}
\examples{
y <- log(AirPassengers)
m1 <- UCmodel(y, model = "llt/eq/arma(0,0)")
tsdiag(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, \code{\link{UCsmooth}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
