% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tsDisplay}
\alias{tsDisplay}
\title{tsDisplay}
\usage{
tsDisplay(y, nCoef = 25, nPar = 0, s = NA)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{nCoef}{number of autocorrelation coefficients to estimate}

\item{nPar}{number of parameters in a model if y is a residual}

\item{s}{seasonal period, number of observations per year}
}
\description{
Displays time series plot with autocorrelation functions
}
\examples{
tsDisplay(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
