% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCestim.R
\name{UCcommand}
\alias{UCcommand}
\title{UCcommand}
\usage{
UCcommand(command, sys)
}
\arguments{
\item{command}{Command to execute: "forecast", "validate", "filter", "smooth", "disturb", "components", "all"}

\item{sys}{A \code{UComp} object created with \code{UC}}
}
\value{
The input \code{UComp} object with the appropriate fields filled in
}
\description{
Auxiliar function for UC modeling
}
\examples{
\donttest{
cycle <- UChp(USgdp)
plot(cycle)
}
}
\seealso{
\code{\link{UC}}, \code{\link{UCforecast}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCcomponents}}, \code{\link{UCdisturb}}
}
\author{
Diego J. Pedregal
}
