% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCestim.R
\name{UCcomponents}
\alias{UCcomponents}
\title{UCcomponents}
\usage{
UCcomponents(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UC} or \code{UCforecast}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\itemize{
\item comp:  Estimated components in matrix form
\item compV: Estimated components variance in matrix form
}
}
\description{
Estimates unobserved components of UC models
Standard methods applicable to UComp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
\dontrun{
m1 <- UC(log(AirPassengers))
m1 <- UCcomponents(m1)
}
}
\seealso{
\code{\link{UC}}, \code{\link{UCforecast}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}},
         \code{\link{UChp}}
}
\author{
Diego J. Pedregal
}
