% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCfilter.R
\name{UCfilter}
\alias{UCfilter}
\title{UCfilter}
\usage{
UCfilter(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UCmodel}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{yFit}{Fitted values of output}
\item{yFitV}{Variance of fitted values of output}
\item{a}{State estimates}
\item{P}{Variance of state estimates}
}
\description{
Runs the Kalman Filter for UC models
}
\examples{
m1 <- UCmodel(log(AirPassengers))
m1 <- UCfilter(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
