% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UComp-package.R
\name{UComp}
\alias{UComp}
\alias{UComp-package}
\title{UComp}
\description{
Package for time series modelling and forecasting of 
times series models inspired on different sources:
}
\details{
\itemize{
\item Unobserved Components models due to A.C. Harvey (Basic Structural Model: BSM), enhanced with automatic identification 
tools by Diego J. Pedregal. 
\item ExponenTial Smoothing by R.J. Hyndman and colaborators.
\item ARIMA models by V. Gómez and A. Maravall
\item Tobit ETS models by Pedregal, Trapero and Holgado
}

The package is designed for automatic identification 
among a wide range of possible models. The models may include exogenous variables. ARMA irregular components 
and automatic detection of outliers in some instances.
}
\section{References}{

Harvey AC (1989). Forecasting, Structural Time Series Models and the Kalman Filter. Cam- bridge University Press.

de Jong, P & Penzer, J (1998). Diagnosing Shocks in Time Series, Journal of the American Statistical Association, 93, 442, 796-806.

Pedregal, DJ, & Young, PC (2002). Statistical approaches to modelling and forecasting time series. In M. Clements,
& D. Hendry (Eds.), Companion to economic forecasting (pp. 69–104). Oxford: Blackwell Publishers.

Durbin J, Koopman SJ (2012). Time Series Analysis by State Space Methods. 38. Oxford University Press.

Proietti T and Luati A (2013). Maximum likelihood estimation of time series models: the Kalman filter and beyond, in Handbook of research methods and applications in empirical macroeconomics, ed. Nigar Hashimzade and Michael Thornton, E. Elgar, UK.

Hyndman RJ, Koehler AB, Ord JK and Snyder RD (2008), Forecasting with exponential smoothing, The State Sapce approach,
Berlin, Springer-Verlag.

Gómez V and Maravall, A (2000), Automatic methods for univariate series. In Peña, D., Tiao, G.C. and Tsay R.S., A course in time series analyis. Wiley.

Trapero JR, Holgado E, Pedregal DJ (2024), Demand forecasting under lost sales stock policies, International Journal of Forecasting, 40, 1055-1068.
}

\section{Maintainer}{
 Diego J. Pedregal
}

\author{
Diego J. Pedregal
}
