% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USgdp}
\alias{USgdp}
\title{US GDP}
\format{
Time series objects.
\describe{
  Quarterly data from 1962 to 2019
}
}
\source{
\href{https://www.nber.org/}{USgdp}
}
\usage{
USgdp
}
\description{
Seasonally adjusted quarterly US real gross domestic product (USgdp).
}
\examples{
\dontrun{
 USgdp
}
}
\keyword{datasets}
