% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{varTest}
\alias{varTest}
\title{varTest}
\usage{
varTest(y, parts = 1/3)
}
\arguments{
\item{y}{a vector, ts or tsibble object}

\item{parts}{portion of sample to estimate variances}
}
\value{
Table with test results
}
\description{
Ratio of variances test
}
\examples{
varTest(AirPassengers)
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{zplane}}, \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
