% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{zplane}
\alias{zplane}
\title{zplane}
\usage{
zplane(MApoly = 1, ARpoly = 1)
}
\arguments{
\item{MApoly}{coefficients of numerator polynomial in descending order}

\item{ARpoly}{coefficients of denominator polynomial in descending order}
}
\description{
Real-imaginary plane to show roots of digital filters (ARMA)
}
\details{
Shows the real-imaginary plane to show zeros (roots of numerator or
MA polynomial) and poles (roots of denominator of AR polynomial). Unit roots
and real vs imaginary roots can be seen by eye
}
\examples{
zplane(c(1, -2, 1), c(1, -0.8))
}
\seealso{
\code{\link{colMedians}}, \code{\link{rowMedians}}, \code{\link{tests}},
         \code{\link{sumStats}}, \code{\link{gaussTest}}, \code{\link{ident}},
         \code{\link{cusum}}, \code{\link{varTest}}, \code{\link{conv}},
         \code{\link{armaFilter}}, \code{\link{dif}}, \code{\link{roots}},
         \code{\link{acft}}, \code{\link{slide}},
         \code{\link{plotSlide}}, \code{\link{Accuracy}}, \code{\link{tsDisplay}},
         \code{\link{size}}
}
\author{
Diego J. Pedregal
}
