% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airpas}
\alias{airpas}
\title{Airpassengers in Spain}
\format{
Time series objects.
\describe{
  Monthly data from January 1992 to December 2019
 \item{airpas}{air passengers data}
  \item{u}{Easter and Busindess day dummy variables}
}
}
\source{
\href{https://www.ine.es}{airpas}
}
\usage{
airpas
}
\description{
Foreign arrivals by air in Spain in thousands of passengers (y).
Easter and Business day dummy variables for airpas data (u).
}
\examples{
\dontrun{
 airpas
}
}
\keyword{datasets}
