% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSvalidate.R
\name{ETSvalidate}
\alias{ETSvalidate}
\title{ETSvalidate}
\usage{
ETSvalidate(m)
}
\arguments{
\item{m}{an object of type \code{UComp} created with \code{UCmodel}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for UC models
}
\examples{
\dontrun{
m1 <- ETSmodel(log(gdp))
m1 <- ETSvalidate(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSmodel}}, \code{\link{ETSvalidate}},
         \code{\link{ETScomponents}}
}
\author{
Diego J. Pedregal
}
