% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Spanish GDP}
\format{
Time series objects.
\describe{
  Quarterly since 1995
}
}
\source{
\href{http://serviciosede.mineco.gob.es/indeco/}{gdp}
}
\usage{
gdp
}
\description{
Quarterly real Spanish Gross Domestic Product (gdp)
}
\examples{
\dontrun{
 gdp
}
}
\keyword{datasets}
