% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OECDgdp}
\alias{OECDgdp}
\title{OECD GDP}
\format{
Time series objects.
\describe{
  Quarterly data from 1962 to 2019
}
}
\source{
\href{http://serviciosede.mineco.gob.es/indeco/}{OECDgdp}
}
\usage{
OECDgdp
}
\description{
Seasonally adjusted quarterly OECD real gross domestic product (OECDgdp).
}
\examples{
\dontrun{
 OECDgdp
}
}
\keyword{datasets}
