% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSS3functions.R, R/PTSS3functions.R
\name{plot.ETS}
\alias{plot.ETS}
\alias{plot.PTS}
\title{plot.ETS}
\usage{
\method{plot}{ETS}(x, ...)

\method{plot}{PTS}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to function.}
}
\description{
Plot components of ETS object

Plot components of PTS object
}
\details{
See help of \code{ETS}.

See help of \code{PTS}.
}
\examples{
\dontrun{
m1 <- ETSmodel(log(gdp))
plot(m1)
}
\dontrun{
m1 <- PTS(log(AirPassengers))
plot(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSmodel}}, \code{\link{ETSvalidate}},
         \code{\link{ETScomponents}}, \code{\link{ETSestim}}

\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
