% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETSS3functions.R, R/PTSS3functions.R
\name{print.ETS}
\alias{print.ETS}
\alias{print.PTS}
\title{print.ETS}
\usage{
\method{print}{ETS}(x, ...)

\method{print}{PTS}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to handle the way to print output.}
}
\description{
Prints an ETS object

Prints a PTS object
}
\details{
See help of \code{ETS}.

See help of \code{PTS}.
}
\examples{
\dontrun{
m1 <- ETSmodel(log(gdp))
print(m1)
}
\dontrun{
m1 <- PTSmodel(log(AirPassengers))
print(m1)
}
}
\seealso{
\code{\link{ETS}}, \code{\link{ETSmodel}}, \code{\link{ETSvalidate}},
         \code{\link{ETScomponents}}, \code{\link{ETSestim}}

\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
