% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCcomponents.R
\name{UCcomponents}
\alias{UCcomponents}
\title{UCcomponents}
\usage{
UCcomponents(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UC}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{comp}{Estimated components in matrix form}
\item{compV}{Estimated components variance in matrix form}
}
\description{
Estimates components of UC models
Standard methods applicable to UComp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
m1 <- UC(log(AirPassengers))
m1 <- UCcomponents(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}},
         \code{\link{UChp}}
}
\author{
Diego J. Pedregal
}
