% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCestim.R
\name{UCestim}
\alias{UCestim}
\title{UCestim}
\usage{
UCestim(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UC}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{p}{Estimated parameters}
\item{v}{Estimated innovations (white noise in correctly specified models)}
\item{yFor}{Forecasted values of output}
\item{yForV}{Variance of forecasted values of output}
\item{criteria}{Value of criteria for estimated model}
}
\description{
Estimates and forecasts UC models
}
\details{
\code{UCestim} estimates and forecasts a time series using an
UC model.
Standard methods applicable to UComp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
m1 <- UCsetup(log(AirPassengers))
m1 <- UCestim(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}}, \code{\link{UCcomponents}},
         \code{\link{UChp}}
}
\author{
Diego J. Pedregal
}
