% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCfilter.R
\name{UChp}
\alias{UChp}
\title{UChp}
\usage{
UChp(y, lambda = 1600)
}
\arguments{
\item{y}{A time series object}

\item{lambda}{Smoothing constant (default: 1600)}
}
\value{
The cycle estimation
}
\description{
Hodrick-Prescott filter estimation
}
\examples{
cycle <- UChp(USgdp)
plot(cycle)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCcomponents}}, \code{\link{UCdisturb}}
}
\author{
Diego J. Pedregal
}
