% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCfilter.R
\name{UCsmooth}
\alias{UCsmooth}
\title{UCsmooth}
\usage{
UCsmooth(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UC}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{yFit}{Fitted values of output}
\item{yFitV}{Variance of fitted values of output}
\item{a}{State estimates}
\item{P}{Variance of state estimates}
}
\description{
Runs the Fixed Interval Smoother for UC models
Standard methods applicable to UComp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
m1 <- UC(log(AirPassengers))
m1 <- UCsmooth(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCvalidate}}, \code{\link{UCfilter}}, 
         \code{\link{UCdisturb}}, \code{\link{UCcomponents}},
         \code{\link{UChp}}
}
\author{
Diego J. Pedregal
}
