% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCvalidate.R
\name{UCvalidate}
\alias{UCvalidate}
\title{UCvalidate}
\usage{
UCvalidate(sys, printScreen = TRUE)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UC}}

\item{printScreen}{print to screen or just return output table}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for UC models.
Equivalent to print or summary.
}
\examples{
m1 <- UC(log(AirPassengers))
m1 <- UCvalidate(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}}, \code{\link{UCcomponents}},
         \code{\link{UChp}}
}
\author{
Diego J. Pedregal
}
