% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCvalidate.R
\name{UCvalidate}
\alias{UCvalidate}
\title{UCvalidate}
\usage{
UCvalidate(sys)
}
\arguments{
\item{sys}{an object of type \code{UComp} created with \code{UCmodel}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\item{table}{Estimation and validation table}
}
\description{
Shows a table of estimation and diagnostics results for UC models
}
\examples{
m1 <- UCmodel(log(AirPassengers))
m1 <- UCvalidate(m1)
}
\seealso{
\code{\link{UC}}, \code{\link{UCmodel}}, \code{\link{UCfilter}}, 
         \code{\link{UCsmooth}}, \code{\link{UCdisturb}}, \code{\link{UCcomponents}}
}
\author{
Diego J. Pedregal
}
